+++ pytensor/link/mlx/dispatch/__init__.py
@@ -6,7 +6,7 @@ import pytensor.link.mlx.dispatch.basic
 import pytensor.link.mlx.dispatch.elemwise
 import pytensor.link.mlx.dispatch.shape
 import pytensor.link.mlx.dispatch.subtensor
-import pytensor.link.mlx.dispatch.core
+import pytensor.link.mlx.dispatch.tensor_basic
 import pytensor.link.mlx.dispatch.signal
 import pytensor.link.mlx.dispatch.signal.conv
 import pytensor.link.mlx.dispatch.blockwise
